requirements.txt
# Syntax Roulette 依赖包

# 核心依赖
gradio>=4.0.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0

# 可选依赖（如果需要WebSocket支持）
websocket-client>=1.6.0
